conversation manager/requirements.txt
flask
pymongo
nltk
python-dotenv
